+++ actix-session/src/lib.rs
@@ -106,15 +106,6 @@ attached to your sessions. You can enable:
   actix-session = { version = "...", features = ["cookie-session"] }
   ```
 
-- a Redis-based backend via [`actix-redis`](https://docs.rs/actix-redis),
-  [`RedisActorSessionStore`], using the `redis-actor-session` feature flag.
-
-  ```toml
-  [dependencies]
-  # ...
-  actix-session = { version = "...", features = ["redis-actor-session"] }
-  ```
-
 - a Redis-based backend via [`redis-rs`](https://docs.rs/redis-rs), [`RedisSessionStore`], using
   the `redis-rs-session` feature flag.