+++ actix-session/src/storage/mod.rs
@@ -11,18 +11,13 @@ pub use self::{
 #[cfg(feature = "cookie-session")]
 mod cookie;
 
-#[cfg(feature = "redis-actor-session")]
-mod redis_actor;
-
 #[cfg(feature = "redis-rs-session")]
 mod redis_rs;
 
-#[cfg(any(feature = "redis-actor-session", feature = "redis-rs-session"))]
+#[cfg(feature = "redis-rs-session")]
 mod utils;
 
 #[cfg(feature = "cookie-session")]
 pub use cookie::CookieSessionStore;
-#[cfg(feature = "redis-actor-session")]
-pub use redis_actor::{RedisActorSessionStore, RedisActorSessionStoreBuilder};
 #[cfg(feature = "redis-rs-session")]
 pub use redis_rs::{RedisSessionStore, RedisSessionStoreBuilder};